+++ navajo/src/dsa.rs
@@ -9,14 +9,14 @@ mod signing_key;
 mod verifier;
 mod verifying_key;
 
+pub(crate) use signing_key::SigningKey;
 pub(crate) use verifying_key::VerifyingKey;
 pub(crate) type Material = SigningKey;
 
 pub use self::signature::Signature;
 pub use algorithm::Algorithm;
-pub use key_info::DsaKeyInfo;
 use key_pair::KeyPair;
 
+pub use key_info::{KeyInfo, KeyringInfo};
 pub use signer::Signer;
-use signing_key::SigningKey;
 pub use verifier::Verifier;