+++ navajo/src/dsa/signing_key.rs
@@ -45,6 +45,7 @@ impl Key<SigningKey> {
     pub(crate) fn pub_id(&self) -> &str {
         &self.material().pub_id
     }
+
     pub(crate) fn verifying_key(&self) -> VerifyingKey {
         self.material().verifying_key.clone()
     }
@@ -110,7 +111,6 @@ impl SigningKey {
     where
         N: Rng,
     {
-        let metadata = metadata.map(Arc::new);
         Self::generate(rng, algorithm, pub_id, metadata)
     }
 
@@ -118,7 +118,7 @@ impl SigningKey {
         algorithm: Algorithm,
         pub_id: String,
         key_pair: KeyPair,
-        metadata: Arc<Metadata>,
+        metadata: Metadata,
     ) -> Result<Self, KeyError> {
         let inner = Arc::new(Inner::from_key_pair(algorithm, &key_pair)?);
         let verifying_key =
@@ -136,7 +136,7 @@ impl SigningKey {
         rng: &N,
         algorithm: Algorithm,
         pub_id: String,
-        metadata: Option<Arc<Metadata>>,
+        metadata: Option<Metadata>,
     ) -> Self
     where
         N: Rng,
@@ -171,7 +171,7 @@ impl<'de> Deserialize<'de> for SigningKey {
             alg: Algorithm,
             pub_id: String,
             #[serde(default)]
-            metadata: Arc<Metadata>,
+            metadata: Metadata,
         }
         let SerializedKey {
             value: key_pair,
@@ -205,7 +205,6 @@ impl KeyMaterial for SigningKey {
     fn algorithm(&self) -> Self::Algorithm {
         self.algorithm
     }
-
     fn kind() -> crate::primitive::Kind {
         Kind::Dsa
     }